refactor: optimiz component & clean code

+++ base-common-starter/base-feign-spring-boot-starter/src/main/java/com/github/liaomengge/base_common/feign/FeignProperties.java
@@ -42,8 +42,8 @@ public class FeignProperties {
         private int maxConnections = 20;
         private long timeToLive = 300L;
         private TimeUnit timeToLiveUnit = TimeUnit.SECONDS;
-        private int connectTimeout = 5000;
-        private int readTimeout = 5000;
-        private int writeTimeout = 5000;
+        private long connectTimeout = 5000;
+        private long readTimeout = 5000;
+        private long writeTimeout = 5000;
     }
 }
