+++ base-common-utils/src/main/java/com/github/liaomengge/base_common/utils/number/LyBigDecimalUtil.java
@@ -13,41 +13,41 @@ public class LyBigDecimalUtil {
 
     private static final int DIV_SCALE = 2;
 
-    public double add(double v1, double v2) {
+    public BigDecimal add(double v1, double v2) {
         BigDecimal b1 = BigDecimal.valueOf(v1);
         BigDecimal b2 = BigDecimal.valueOf(v2);
-        return b1.add(b2).doubleValue();
+        return b1.add(b2);
     }
 
-    public double sub(double v1, double v2) {
+    public BigDecimal sub(double v1, double v2) {
         BigDecimal b1 = BigDecimal.valueOf(v1);
         BigDecimal b2 = BigDecimal.valueOf(v2);
-        return b1.subtract(b2).doubleValue();
+        return b1.subtract(b2);
     }
 
-    public double mul(double v1, double v2) {
+    public BigDecimal mul(double v1, double v2) {
         BigDecimal b1 = BigDecimal.valueOf(v1);
         BigDecimal b2 = BigDecimal.valueOf(v2);
-        return b1.multiply(b2).doubleValue();
+        return b1.multiply(b2);
     }
 
-    public double div(double v1, double v2) {
+    public BigDecimal div(double v1, double v2) {
         return div(v1, v2, DIV_SCALE);
     }
 
-    public double div(double v1, double v2, int scale) {
+    public BigDecimal div(double v1, double v2, int scale) {
         if (scale < 0) {
             throw new IllegalArgumentException("The scale must be a positive integer or zero");
         }
         BigDecimal b1 = BigDecimal.valueOf(v1);
         BigDecimal b2 = BigDecimal.valueOf(v2);
-        return b1.divide(b2, scale, BigDecimal.ROUND_HALF_UP).doubleValue();
+        return b1.divide(b2, scale, BigDecimal.ROUND_HALF_UP);
     }
 
-    public double round(double v, int scale) {
+    public BigDecimal round(double v, int scale) {
         if (scale < 0) {
             throw new IllegalArgumentException("The scale must be a positive integer or zero");
         }
-        return BigDecimal.valueOf(v).setScale(scale, RoundingMode.HALF_UP).doubleValue();
+        return BigDecimal.valueOf(v).setScale(scale, RoundingMode.HALF_UP);
     }
 }
